First Deployment Version

+++ SLIDE-x-BENCH/KERNEL/astar/frst.c
@@ -2,8 +2,8 @@
 #include <values.h>
 #include <stdio.h>
 
-typedef int8_t TARGET_TYPE;
-typedef uint8_t TARGET_INDEX;
+typedef double TARGET_TYPE;
+typedef uint64_t TARGET_INDEX;
 
 //void prototype(int8_t size, int8_t a[size][size], int8_t goal);
