+++ SLIDE-x-BENCH/RECIPE/prime/frst.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef float TARGET_TYPE;
+typedef int64_t TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
   TARGET_TYPE i;
